Refactors RackBuilder to fix Rubocop ClassLength cop
